parser is working.  It's now able to convert the json data to the WeatherData object correctly.

+++ assignments/WeatherServiceApp_P2-Assignment3/src/com/example/weatherserviceapp/json/Sys.java
@@ -8,8 +8,8 @@ public class Sys {
 
     private Double message;
     private String country;
-    private Integer sunrise;
-    private Integer sunset;
+    private Long sunrise;
+    private Long sunset;
     private Map<String, Object> additionalProperties = new HashMap<String, Object>();
 
     /**
@@ -63,7 +63,7 @@ public class Sys {
      * @return
      *     The sunrise
      */
-    public Integer getSunrise() {
+    public Long getSunrise() {
         return sunrise;
     }
 
@@ -72,11 +72,11 @@ public class Sys {
      * @param sunrise
      *     The sunrise
      */
-    public void setSunrise(Integer sunrise) {
+    public void setSunrise(Long sunrise) {
         this.sunrise = sunrise;
     }
 
-    public Sys withSunrise(Integer sunrise) {
+    public Sys withSunrise(Long sunrise) {
         this.sunrise = sunrise;
         return this;
     }
@@ -86,7 +86,7 @@ public class Sys {
      * @return
      *     The sunset
      */
-    public Integer getSunset() {
+    public Long getSunset() {
         return sunset;
     }
 
@@ -95,11 +95,11 @@ public class Sys {
      * @param sunset
      *     The sunset
      */
-    public void setSunset(Integer sunset) {
+    public void setSunset(Long sunset) {
         this.sunset = sunset;
     }
 
-    public Sys withSunset(Integer sunset) {
+    public Sys withSunset(Long sunset) {
         this.sunset = sunset;
         return this;
     }
